% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMulti-internal.R
\name{agg}
\alias{agg}
\title{Extract School-Level Covariates}
\usage{
agg(x)
}
\arguments{
\item{x}{a vector containing student-level observations for a school. If it
is a factor it must contain only a single level.}
}
\value{
A single value of the same type as the input vector.
}
\description{
Given a vector of variables of interest for students in a single school,
extracts a single value for the school
}
\details{
If the input is numeric, \code{agg} returns the mean; if the input is not
numeric, an error will be thrown unless all values are the same, in which
case the single unique value will be returned.
}
\author{
Luke Keele, Penn State University, \email{ljk20@psu.edu}

Sam Pimentel, University of Pennsylvania, \email{spi@wharton.upenn.edu}
}
\keyword{internal}
