% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceTable.R
\name{balanceTable}
\alias{balanceTable}
\title{Create Balance Table}
\usage{
balanceTable(
  df.orig,
  df.match = NULL,
  treatment,
  school.id = NULL,
  var.names = NULL,
  treat.wts = NULL,
  ctrl.wts = NULL,
  mt.wts = NULL,
  mc.wts = NULL,
  include.tests = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df.orig}{a data frame containing the data before matching}

\item{df.match}{an optional data frame containing the matched sample. Must
have all variable names to be balanced.}

\item{treatment}{name of the binary indicator for treatment status}

\item{school.id}{Identifier for groups (for example schools); need to pass if
p-values for balance statistics are desired.}

\item{var.names}{List of variable names to calculate balance for.  If NULL,
use all variables found in the df.orig data.frame.}

\item{treat.wts}{optional weights for treated units in the original sample}

\item{ctrl.wts}{optional weights for control units in the original sample}

\item{mt.wts}{optional weights for treated units in the matched sample}

\item{mc.wts}{optional weights for treated units in the matched sample}

\item{include.tests}{Include tests of imbalance on covariates (TRUE/FALSE).}

\item{verbose}{a logical value indicating whether detailed output should be
printed.}
}
\value{
A data.frame of balance measures, with one row for each covariate in
  \code{df.orig} except \code{treatment}, and columns for treated and control
  means, standardized differences in means, p-values from a 2-sample t-test,
  and p-values from either Fisher's exact test (if the covariate is binary)
  or a Wilcoxon signed rank test otherwise.  If \code{df.match} is specified
  there are twice as many columns, one set for the pre-match samples and one
  set for the post-match samples.
}
\description{
Given an unmatched sample of treated and control units and (optionally) a
matched sample from the same data, produces a table with pre- and post-match
measures of covariate balance.
}
\references{
Rosenbaum, Paul R. (2002). \emph{Observational Studies}.
  Springer-Verlag.

  Rosenbaum, Paul R. (2010). \emph{Design of Observational Studies}.
  Springer-Verlag.
}
\keyword{internal}
