% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMultiResult.R
\name{describe_data_counts}
\alias{describe_data_counts}
\title{Describe the count of students and schools}
\usage{
describe_data_counts(data, school.id, treatment)
}
\arguments{
\item{data}{Dataset (student level)}

\item{school.id}{String name of ID column in data (the grouping variable)}

\item{treatment}{String name of the treatment variable.}
}
\value{
List of two things: school and student counts.
}
\description{
Given a school ID and treatment variable, count up number of schools and
students.
}
\author{
Luke Miratrix
}
