% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMulti-internal.R
\name{is.binary}
\alias{is.binary}
\title{Check if a variable is binary}
\usage{
is.binary(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
a logical value, \code{TRUE} if the vector contains only 0s and 1s
and \code{FALSE} otherwise.
}
\description{
Examines a vector that is not coded as a logical to see if it contains only
0s and 1s.
}
\author{
Luke Keele, Penn State University, \email{ljk20@psu.edu}

Sam Pimentel, University of Pennsylvania, \email{spi@wharton.upenn.edu}
}
\keyword{internal}
