% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMultiResult.R
\name{matchMultiResult}
\alias{matchMultiResult}
\alias{is.matchMultiResult}
\alias{print.matchMultiResult}
\alias{summary.matchMultiResult}
\title{matchMultiResult object for results of power calculations}
\usage{
is.matchMultiResult(x)

\method{print}{matchMultiResult}(x, n = 10, ...)

\method{summary}{matchMultiResult}(object, ...)
}
\arguments{
\item{x}{a matchMultiResult object (except for is.matchMultiResult, where it is a generic
object to check).}

\item{n}{Number of lines of search path to print, max.}

\item{...}{Extra options passed to print.matchMultiResult}

\item{object}{Object to summarize.}

\item{header}{FALSE means skip some header info on the result, just print
the data.frame of actual results.}

\item{search}{FALSE means don't print the search path for a result for
mdes or sample.}
}
\value{
is.matchMultiResult: TRUE if object is a matchMultiResult object.
}
\description{
The matchMultiResult object is an S3 class that holds the results from the matchMulti call.

matchMulti result objects have the matched datasets inside of them.
}
