% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMulti-package.R
\docType{data}
\name{minischool}
\alias{minischool}
\title{Mini-data set for illustration}
\format{
A data frame with 1500 rows and 12 variables, as described in the
  `catholic_schools` dataset.
}
\source{
See documentation page for `catholic_schools` dataset.
}
\description{
The Catholic schools dataset subset to a smaller number of
  schools (with only 6 Catholic schools).  See full dataset documentation for
  more information.
}
\seealso{
catholic_schools
}
