% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMulti-internal.R
\name{pe_func}
\alias{pe_func}
\title{Outcome analysis.}
\usage{
pe_func(beta, obj, out.name = NULL, schl_id_name = NULL, treat.name = NULL)
}
\arguments{
\item{beta}{Point estimate value}

\item{obj}{A multiMatch object}

\item{out.name}{Name of outcome covariate}

\item{schl_id_name}{Name of school (group) identifier}

\item{treat.name}{Name of treatment indicator}
}
\value{
A point estimate for constant-additive treatment effect.
}
\description{
Calculates Hodges-Lehmann point estimate for treatment effect.
}
\references{
Rosenbaum, Paul R. (2002). \emph{Observational Studies}.
Springer-Verlag.

Rosenbaum, Paul R. (2010). \emph{Design of Observational Studies}.
Springer-Verlag.
}
\author{
Luke Keele, Penn State University, \email{ljk20@psu.edu}

Sam Pimentel, University of Pennsylvania, \email{spi@wharton.upenn.edu}
}
\keyword{internal}
