\name{is.binary}
\alias{is.binary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Check if a variable is binary
}
\description{
Examines a vector that is not coded as a logical to see if it contains only 0s and 1s.
}
\usage{
is.binary(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
	A vector.
}
}

\value{
a logical value, \code{TRUE} if the vector contains only 0s and 1s and \code{FALSE} otherwise.
} 

\author{
  Luke Keele, Penn State University, \email{ljk20@psu.edu}

  Sam Pimentel, University of Pennsylvania, \email{spi@wharton.upenn.edu}
}

\keyword{internal}
