% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMulti-internal.R
\name{resolve.cols}
\alias{resolve.cols}
\title{Ensure Dataframes Share Same Set Columns}
\usage{
resolve.cols(df1, df2)
}
\arguments{
\item{df1}{a dataframe.}

\item{df2}{a dataframe.}
}
\value{
a dataframe
}
\description{
Takes in two dataframes.  For each column name that is in the second frame
but not in the first frame, a new column of zeroes is added to the first
frame.
}
\author{
Luke Keele, Penn State University, \email{ljk20@psu.edu}

Sam Pimentel, University of Pennsylvania, \email{spi@wharton.upenn.edu}
}
\keyword{internal}
