% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMultiResult.R
\name{tally_schools}
\alias{tally_schools}
\title{Tally schools and students in a given dataset}
\usage{
tally_schools(data, school.id, treatment)
}
\arguments{
\item{data}{Dataset (student level)}

\item{school.id}{String name of ID column in data (the grouping variable)}

\item{treatment}{String name of the treatment variable.}
}
\value{
List of two things: school and student counts (invisible).
}
\description{
Returns a count of schools, without printing
anything.
}
\seealso{
describe_data_counts
}
\author{
Luke Miratrix
}
